Attempt at a bug fix in set_pot
Bug description: After converging the potential and completing the first geometry optimization loop, the constant potential relax routine gets stuck in a loop and creates tons of iter*.traj files. This is caused by the calculator object not being updated by set_pot since the potential is still at the set point, and so the atoms.get_potential_energy() call in const_U_relax does not execute VASP. Added a catch in set_pot to hopefully fix this.