Updates to two functions and a typo fix
In the scale_metal function: multiply atom positions by the ratio of lattice constants rather than trying to translate atoms.

In the get_zval function: fixed a typo in one of the values.

In const_u_dimer: add `iopt' as an input variable (will set to 2 by default, compared to vasp which sets 1 by default) and add a 'bug fix' (not actually fixed) to circumvent looping through the main loop too fast.