Update to set_pot function
Fixing a bug in set_pot and making it a bit more robust. There are several checks in place to ensure the routine will not get stuck in an infinite loop by forcing the number of electrons to change if there was no change in the previous step.

In addition, the initial guess for the capacitance now takes into account varying cell areas. Before it used an initial guess of 1 e/V, which is a good guess for cells around 100 Å^2. Updated version scales this capacitance by the area of the current cell.

Finally, if the data is available, the capacitance is now estimated from three data points (in U vs q) rather than the most recent two. Especially helpful if the previous step in NELECT was very small.