Fixed a consistency issue with scale_metal
Previously, old_sym was a required keyword when really it is only needed if new_sym is also provided, which changes not only the lattice constant but also the element. This change makes old_sym optional and adds an assertion to require it if new_sym is in fact provided.